+++ src/change/sdp.rs
@@ -22,7 +22,7 @@ use crate::RtcError;
 use crate::{Candidate, IceCreds};
 
 pub use crate::sdp::{SdpAnswer, SdpOffer};
-use crate::streams::{Streams, DEFAULT_RTX_CACHE_DROP_RATIO, DEFAULT_RTX_CACHE_DURATION};
+use crate::streams::{Streams, DEFAULT_RTX_CACHE_DURATION, DEFAULT_RTX_RATIO_CAP};
 
 /// Changes to the Rtc via SDP Offer/Answer dance.
 pub struct SdpApi<'a> {
@@ -862,11 +862,7 @@ fn ensure_stream_tx(session: &mut Session) {
                 session.send_buffer_video
             };
 
-            stream.set_rtx_cache(
-                size,
-                DEFAULT_RTX_CACHE_DURATION,
-                DEFAULT_RTX_CACHE_DROP_RATIO,
-            );
+            stream.set_rtx_cache(size, DEFAULT_RTX_CACHE_DURATION, DEFAULT_RTX_RATIO_CAP);
         }
     }
 }
@@ -902,11 +898,7 @@ fn add_pending_changes(session: &mut Session, pending: Changes) {
                 session.send_buffer_video
             };
 
-            stream.set_rtx_cache(
-                size,
-                DEFAULT_RTX_CACHE_DURATION,
-                DEFAULT_RTX_CACHE_DROP_RATIO,
-            );
+            stream.set_rtx_cache(size, DEFAULT_RTX_CACHE_DURATION, DEFAULT_RTX_RATIO_CAP);
         }
     }
 }