+++ CHANGELOG.md
@@ -9,6 +9,7 @@
   * Fix bug using unreliable channels by default #548
   * New add_channel_with_config() for configured data channels #548
   * Fix RTX stops working after packet loss spike #566
+  * Configure RTX ratio cap via `StreamTx::set_rtx_cache` #570
 
 # 0.6.1
   * Force openssl to be >=0.10.66 #545